uskat.s5:+branch: "db_storage_format_(mariadb_only,legacy)"

+++ org.toxsoft.uskat.s5.histdata10/src/org/toxsoft/uskat/s5/server/backend/supports/histdata/impl/sequences/async/S5HistDataAsyncFloatingBlobEntity1.java
@@ -15,7 +15,7 @@ import jakarta.persistence.*;
  */
 @Entity
 public class S5HistDataAsyncFloatingBlobEntity1
-    extends S5SequenceAsyncBlob<S5HistDataAsyncFloatingEntity1, double[], byte[]> {
+    extends S5SequenceAsyncBlob<S5HistDataAsyncFloatingEntity1, double[], double[]> {
 
   private static final long serialVersionUID = 157157L;
