+++ org.toxsoft.uskat.s5.histdata10/src/org/toxsoft/uskat/s5/server/backend/supports/histdata/impl/sequences/async/S5HistDataAsyncIntegerBlobEntity5.java
@@ -15,7 +15,7 @@ import jakarta.persistence.*;
  */
 @Entity
 public class S5HistDataAsyncIntegerBlobEntity5
-    extends S5SequenceAsyncBlob<S5HistDataAsyncIntegerEntity5, long[], byte[]> {
+    extends S5SequenceAsyncBlob<S5HistDataAsyncIntegerEntity5, long[], long[]> {
 
   private static final long serialVersionUID = 157157L;
