+++ org.toxsoft.uskat.s5/src/org/toxsoft/uskat/s5/server/backend/supports/histdata/impl/sequences/async/S5HistDataAsyncTimestampBlobEntity.java
@@ -15,7 +15,7 @@ import jakarta.persistence.*;
  */
 @Entity
 public class S5HistDataAsyncTimestampBlobEntity
-    extends S5SequenceAsyncBlob<S5HistDataAsyncTimestampEntity, long[], byte[]> {
+    extends S5SequenceAsyncBlob<S5HistDataAsyncTimestampEntity, long[], long[]> {
 
   private static final long serialVersionUID = 157157L;
